Consolidate all Status in core.
Status get compared with each other and we likely want to consolidate
all the status.
